feat: TableInterfaceTrait is added
